Add RemoveNoteHandler

+++ nvi-commons/src/main/java/no/sikt/nva/nvi/common/service/NviService.java
@@ -10,7 +10,6 @@ import java.util.UUID;
 import no.sikt.nva.nvi.common.db.Candidate;
 import no.sikt.nva.nvi.common.db.NviCandidateRepository;
 import no.sikt.nva.nvi.common.db.NviPeriodRepository;
-
 import no.sikt.nva.nvi.common.db.model.DbApprovalStatus;
 import no.sikt.nva.nvi.common.db.model.DbCandidate;
 import no.sikt.nva.nvi.common.db.model.DbInstitutionPoints;
@@ -88,9 +87,13 @@ public class NviService {
         return nviCandidateRepository.findById(identifier).orElseThrow();
     }
 
-    public Candidate deleteNote(UUID candidateIdentifier, UUID noteIdentifier) {
-        nviCandidateRepository.deleteNote(candidateIdentifier, noteIdentifier);
-        return nviCandidateRepository.findById(candidateIdentifier).orElseThrow();
+    public Candidate deleteNote(UUID candidateIdentifier, UUID noteIdentifier, String username) {
+        DbNote note = nviCandidateRepository.getNoteById(candidateIdentifier, noteIdentifier);
+        if (note.user().value().equals(username)) {
+            nviCandidateRepository.deleteNote(candidateIdentifier, noteIdentifier);
+            return nviCandidateRepository.findById(candidateIdentifier).orElseThrow();
+        }
+        throw new IllegalArgumentException("User not allowed to remove others note.");
     }
 
     private static boolean isInteger(String value) {
@@ -132,6 +135,18 @@ public class NviService {
                    .build();
     }
 
+    private static DbApprovalStatus updateStatus(DbApprovalStatus oldApprovalStatus,
+                                                 DbApprovalStatus newApprovalStatus) {
+        return switch (newApprovalStatus.status()) {
+            case APPROVED, REJECTED -> finalizeStatus(oldApprovalStatus, newApprovalStatus);
+            case PENDING -> resetStatus(oldApprovalStatus);
+        };
+    }
+
+    private static boolean updateIsAssignee(DbApprovalStatus oldApprovalStatus, DbApprovalStatus newApprovalStatus) {
+        return oldApprovalStatus.status().equals(newApprovalStatus.status());
+    }
+
     private DbNviPeriod injectCreatedBy(DbNviPeriod period) {
         return period.copy()
                    .createdBy(getPeriod(period.publishingYear()).createdBy())
@@ -157,18 +172,6 @@ public class NviService {
         } else {
             return updateStatus(oldApprovalStatus, newApprovalStatus);
         }
-    }
-
-    private static DbApprovalStatus updateStatus(DbApprovalStatus oldApprovalStatus,
-                                                        DbApprovalStatus newApprovalStatus) {
-        return switch (newApprovalStatus.status()) {
-            case APPROVED, REJECTED -> finalizeStatus(oldApprovalStatus, newApprovalStatus);
-            case PENDING -> resetStatus(oldApprovalStatus);
-        };
-    }
-
-    private static boolean updateIsAssignee(DbApprovalStatus oldApprovalStatus, DbApprovalStatus newApprovalStatus) {
-        return oldApprovalStatus.status().equals(newApprovalStatus.status());
     }
 
     private DbApprovalStatus updateAssignee(DbApprovalStatus oldApprovalStatus, DbApprovalStatus newApprovalStatus) {
